reorganized code
